requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alpha_vantage==3.0.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.66.0
anyio==4.10.0
attrs==25.3.0
banks==2.2.0
beautifulsoup4==4.13.5
black==25.1.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cssselect==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
faiss-cpu==1.12.0
fastapi==0.116.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.19.1
filetype==1.2.0
flake8==7.3.0
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.75
langchain-text-splitters==0.3.11
langsmith==0.4.27
llama-cloud==0.1.35
llama-index==0.10.43
llama-index-agent-openai==0.1.7
llama-index-cli==0.1.13
llama-index-core==0.10.43
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-instrumentation==0.4.0
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.1.26
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-index-vector-stores-faiss==0.1.2
llama-index-workflows==1.3.0
llama-parse==0.4.9
llamaindex-py-client==0.1.19
lxml==6.0.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mccabe==0.7.0
mpmath==1.3.0
multidict==6.6.4
mypy==1.17.1
mypy_extensions==1.1.0
narwhals==2.4.0
neo4j==5.28.2
nest-asyncio==1.6.0
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
openai==1.107.1
orjson==3.11.3
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
pluggy==1.6.0
propcache==0.3.2
protobuf==6.32.0
pyarrow==21.0.0
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyflakes==3.4.0
Pygments==2.19.2
PyMuPDF==1.26.4
pyparsing==3.2.3
pypdf==4.3.1
pytest==8.4.2
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.47.3
streamlit==1.49.1
striprtf==0.0.26
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1
zstandard==0.24.0
